projects/cli based Poetry Analyzer/pyproject.toml
[project]
name = "cli-based-poetry-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
]
